Split the view of one Eye into its separate file called Eye.vue
the handler for the mouse event is on the svg with @mousemove

- What doesn't work:
--------------------
the mouse movement is only detected when the mouse is on the svg itself.

the handling of the id isnt working right. the combined ids (i.e: ${id}_closeLod in svg) are not recognized.

the styling of the eye isnt right, they overlap